Add dashboard after login
