--- app/routes/join.tsx
@@ -1,169 +1,0 @@
-import type { ActionFunction, LoaderArgs, MetaFunction } from "@remix-run/node";
-import { json, redirect } from "@remix-run/node";
-import { Form, Link, useActionData, useSearchParams } from "@remix-run/react";
-import { createUserSession, getUserId } from "~/server/session.server";
-import { createUser } from "~/server/models/user.server";
-import { validateEmail } from "~/utils";
-import * as React from "react";
-import invariant from "tiny-invariant";
-
-export const meta: MetaFunction = () => {
-  return {
-    title: "Sign Up",
-  };
-};
-
-interface ActionData {
-  errors: {
-    email?: string;
-    password?: string;
-  };
-}
-
-export async function loader({ request }: LoaderArgs) {
-  const userId = await getUserId(request);
-  if (userId) return redirect("/");
-  return json({});
-}
-
-export const action: ActionFunction = async ({ request }) => {
-  const formData = await request.formData();
-  const email = formData.get("email");
-  const password = formData.get("password");
-  const redirectTo = formData.get("redirectTo");
-
-  // Ensure the email is valid
-  if (!validateEmail(email)) {
-    return json<ActionData>(
-      { errors: { email: "Email is invalid." } },
-      { status: 400 }
-    );
-  }
-
-  // What if a user sends us a password through other means than our form?
-  if (typeof password !== "string") {
-    return json(
-      { errors: { password: "Valid password is required." } },
-      { status: 400 }
-    );
-  }
-
-  // Enforce minimum password length
-  if (password.length < 6) {
-    return json<ActionData>(
-      { errors: { password: "Password is too short." } },
-      { status: 400 }
-    );
-  }
-
-  // A user could potentially already exist within our system
-  // and we should communicate that well
-  // const existingUser = await getProfileByEmail(email);
-  // if (existingUser) {
-  //   return json<ActionData>(
-  //     { errors: { email: "A user already exists with this email." } },
-  //     { status: 400 }
-  //   );
-  // }
-
-  const user = await createUser(email, password);
-  invariant(user, "User not found!!");
-  return createUserSession({
-    request,
-    userId: user.id,
-    remember: false,
-    redirectTo: typeof redirectTo === "string" ? redirectTo : "/",
-  });
-};
-
-export default function Join() {
-  const [searchParams] = useSearchParams();
-  const redirectTo = searchParams.get("redirectTo") ?? undefined;
-
-  const actionData = useActionData() as ActionData;
-  const emailRef = React.useRef<HTMLInputElement>(null);
-  const passwordRef = React.useRef<HTMLInputElement>(null);
-
-  React.useEffect(() => {
-    if (actionData?.errors?.email) {
-      emailRef?.current?.focus();
-    }
-
-    if (actionData?.errors?.password) {
-      passwordRef?.current?.focus();
-    }
-  }, [actionData]);
-
-  return (
-    <div className="flex min-h-full flex-col justify-center">
-      <div className="mx-auto w-full max-w-md px-8">
-        <Form className="space-y-6" method="post" noValidate>
-          <div>
-            <label className="text-sm font-medium" htmlFor="email">
-              <span className="block text-gray-700">Email Address</span>
-              {actionData?.errors?.email && (
-                <span className="block pt-1 text-red-700" id="email-error">
-                  {actionData?.errors?.email}
-                </span>
-              )}
-            </label>
-            <input
-              className="w-full rounded border border-gray-500 px-2 py-1 text-lg"
-              type="email"
-              name="email"
-              id="email"
-              required
-              aria-invalid={actionData?.errors?.email ? true : undefined}
-              aria-describedby="email-error"
-              ref={emailRef}
-            />
-          </div>
-          <div>
-            <label className="text-sm font-medium" htmlFor="password">
-              <span className="block text-gray-700">Password</span>
-              <span className="block font-light text-gray-700">
-                Must have at least 6 characters.
-              </span>
-              {actionData?.errors?.password && (
-                <span className="pt-1 text-red-700" id="password-error">
-                  {actionData?.errors?.password}
-                </span>
-              )}
-            </label>
-            <input
-              id="password"
-              type="password"
-              name="password"
-              className="w-full rounded border border-gray-500 px-2 py-1 text-lg"
-              autoComplete="new-password"
-              aria-invalid={actionData?.errors?.password ? true : undefined}
-              aria-describedby="password-error"
-              ref={passwordRef}
-            />
-          </div>
-          <button
-            className="w-full rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
-            type="submit"
-          >
-            Create Account
-          </button>
-          <input type="hidden" name="redirectTo" value={redirectTo} />
-          <div className="flex items-center justify-center">
-            <div className="text-center text-sm text-gray-500">
-              Already have an account?{" "}
-              <Link
-                className="text-blue-500 underline"
-                to={{
-                  pathname: "/login",
-                  search: searchParams.toString(),
-                }}
-              >
-                Log in
-              </Link>
-            </div>
-          </div>
-        </Form>
-      </div>
-    </div>
-  );
-}
