+++ app/server/session.server.ts
@@ -29,7 +29,6 @@ export async function getSession(request: Request) {
 export async function getUserId(request: Request) {
   const session = await getSession(request);
   const userId = session.get(USER_SESSION_KEY);
-
   return userId;
 }
